// apps/frontend/src/api/model/assignmentsAssignmentStatus.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * My Catatan Kuliah
 * My Catatan Kuliah API
 * OpenAPI spec version: 1.0.0
 */

export type AssignmentsAssignmentStatus = typeof AssignmentsAssignmentStatus[keyof typeof AssignmentsAssignmentStatus];


export const AssignmentsAssignmentStatus = {
  Todo: 'Todo',
  InProgress: 'InProgress',
  Done: 'Done',
} as const;